Go ahead and remove all pre-v6 code
- Rip off the band aid and remove all v6 code.
- Make further changes to support v6.
