+++ ouroboros/chainsync/byron.go
@@ -1,4 +1,4 @@
-// Copyright 2021 Matt Ho
+// Copyright 2023 SundaeSwap Labs
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,47 +12,77 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Byron block code for Ogmios v6.
+// https://github.com/CardanoSolutions/ogmios/blob/d2b1d70ab5e676b5d053817d57ea8220f2b07317/docs/static/api/specification.yaml
+
 package chainsync
 
 import (
 	"encoding/json"
 )
 
-type ByronBlock struct {
-	Body   ByronBody
-	Hash   string
-	Header ByronHeader
+// BFT Block Root
+type ByronBlockBFT struct {
+	Type                    string             `json:"type,omitempty"`
+	Era                     string             `json:"era,omitempty"`
+	ID                      string             `json:"id,omitempty"`
+	Ancestor                string             `json:"ancestor,omitempty"`
+	Height                  uint64             `json:"height,omitempty"`
+	Slot                    uint64             `json:"slot,omitempty"`
+	Size                    BlockSize          `json:"size,omitempty"`
+	Transactions            []Tx               `json:"transactions,omitempty"`
+	OperationalCertificates []json.RawMessage  `json:"operationalCertificates,omitempty"`
+	Protocol                ByronProtocol      `json:"protocol,omitempty"`
+	Issuer                  ByronBlockIssuer   `json:"issuer,omitempty"`
+	Delegate                ByronBlockDelegate `json:"delegate,omitempty"`
 }
 
-type ByronBody struct {
-	DlgPayload    []json.RawMessage `json:"dlgPayload,omitempty"`
-	TxPayload     []ByronTxPayload  `json:"txPayload,omitempty"`
-	UpdatePayload json.RawMessage   `json:"updatePayload,omitempty"`
+// EBB Block Type
+type ByronBlockEBB struct {
+	Type     string `json:"type,omitempty"`
+	Era      string `json:"era,omitempty"`
+	ID       string `json:"id,omitempty"`
+	Ancestor string `json:"ancestor,omitempty"`
+	Height   uint64 `json:"height,omitempty"`
 }
 
-type ByronHeader struct {
-	BlockHeight     uint64 `json:"blockHeight,omitempty"`
-	GenesisKey      string
-	Epoch           uint32
-	Proof           json.RawMessage
-	PrevHash        string
-	ProtocolMagicId uint64
-	ProtocolVersion ProtocolVersion
-	Signature       json.RawMessage
-	Slot            uint64 `json:"slot,omitempty"`
-	SoftwareVersion map[string]interface{}
+type ByronBlockDelegate struct {
+	VerificationKey string
 }
 
-type ByronTxBody struct {
-	Inputs  []TxIn
-	Outputs []TxOut
+type ByronBlockIssuer struct {
+	VerificationKey string
 }
 
-type ByronTxPayload struct {
-	ID      string
-	Witness []ByronWitness
+type ByronProtocol struct {
+	Version  ProtocolVersion
+	Id       uint64 // aka magic
+	Software map[string]interface{}
+	Update   json.RawMessage
 }
 
-type ByronWitness struct {
-	RedeemWitness map[string]string
+type ResultByronEBB struct {
+	Direction string         `json:"direction,omitempty" dynamodbav:"direction,omitempty"`
+	Block     *ByronBlockEBB `json:"block,omitempty"     dynamodbav:"block,omitempty"`
+	Tip       *Tip           `json:"tip,omitempty"       dynamodbav:"tip,omitempty"`
+}
+
+type ResultByronBFT struct {
+	Direction string         `json:"direction,omitempty" dynamodbav:"direction,omitempty"`
+	Block     *ByronBlockBFT `json:"block,omitempty"     dynamodbav:"block,omitempty"`
+	Tip       *Tip           `json:"tip,omitempty"       dynamodbav:"tip,omitempty"`
+}
+
+type ResponseByronEBB struct {
+	JsonRpc string          `json:"jsonrpc,omitempty" dynamodbav:"jsonrpc,omitempty"`
+	Method  string          `json:"method,omitempty"  dynamodbav:"method,omitempty"`
+	Result  *ResultByronEBB `json:"result,omitempty"  dynamodbav:"result,omitempty"`
+	ID      json.RawMessage `json:"id,omitempty"      dynamodbav:"id,omitempty"`
+}
+
+type ResponseByronBFT struct {
+	JsonRpc string          `json:"jsonrpc,omitempty" dynamodbav:"jsonrpc,omitempty"`
+	Method  string          `json:"method,omitempty"  dynamodbav:"method,omitempty"`
+	Result  *ResultByronBFT `json:"result,omitempty"  dynamodbav:"result,omitempty"`
+	ID      json.RawMessage `json:"id,omitempty"      dynamodbav:"id,omitempty"`
 }
